Connect to postgres
Embedded databases are good for testing but a development and production
environment should use persistent storage. We have decided to use postgres.
The application talks to the postgres database in development profile using
admin role. Currently, DDL and DML are generated by hibernate and import.sql
respectively. In the future we should try to use database migration tool like
liquibase.